refactor: 차 이름 문자열 포장을 위한 CarName 추가
