+++ src/main/java/racingcar/domain/Car.java
@@ -1,42 +1,14 @@
 package racingcar.domain;
 
 public class Car {
-    private final String name;
+    private final CarName name;
     private int position;
-    private MovingPolicy movingPolicy;
+    private final MovingPolicy movingPolicy;
 
     public Car(String name, MovingPolicy movingPolicy) {
-        validateNull(name);
-        validateEmpty(name);
-        validateBlank(name);
-        validateNameLength(name);
-        this.name = name;
+        this.name = new CarName(name);
         this.position = 0;
         this.movingPolicy = movingPolicy;
-    }
-
-    private void validateNull(String name) {
-        if (name == null) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateEmpty(String name) {
-        if (name.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateBlank(String name) {
-        if (name.isBlank()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void validateNameLength(String name) {
-        if (name.length() > 5) {
-            throw new IllegalArgumentException();
-        }
     }
 
     public void move() {
@@ -46,7 +18,7 @@ public class Car {
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public int getPosition() {
